+++ packages/flutter/lib/src/services/binding.dart
@@ -27,11 +27,15 @@ mixin ServicesBinding on BindingBase {
     window.onPlatformMessage = defaultBinaryMessenger.handlePlatformMessage;
     initLicenses();
     SystemChannels.system.setMessageHandler(handleSystemMessage);
+    SystemChannels.lifecycle.setMessageHandler(_handleLifecycleMessage);
+    readInitialLifecycleStateFromNativeWindow();
   }
 
   /// The current [ServicesBinding], if one has been created.
   static ServicesBinding get instance => _instance;
   static ServicesBinding _instance;
+
+  // Platform Messages
 
   /// The default instance of [BinaryMessenger].
   ///
@@ -74,6 +78,8 @@ mixin ServicesBinding on BindingBase {
     }
     return;
   }
+
+  // Licenses
 
   /// Adds relevant licenses to the [LicenseRegistry].
   ///
@@ -161,6 +167,67 @@ mixin ServicesBinding on BindingBase {
   @mustCallSuper
   void evict(String asset) {
     rootBundle.evict(asset);
+  }
+
+  // App life cycle
+
+  /// Whether the application is visible, and if so, whether it is currently
+  /// interactive.
+  ///
+  /// This is set by [handleAppLifecycleStateChanged] when the
+  /// [SystemChannels.lifecycle] notification is dispatched.
+  ///
+  /// The preferred way to watch for changes to this value is using
+  /// [WidgetsBindingObserver.didChangeAppLifecycleState].
+  ui.AppLifecycleState get lifecycleState => _lifecycleState;
+  ui.AppLifecycleState _lifecycleState;
+
+  /// Called when the application lifecycle state changes.
+  ///
+  /// Notifies all the observers using
+  /// [WidgetsBindingObserver.didChangeAppLifecycleState].
+  ///
+  /// This method exposes notifications from [SystemChannels.lifecycle].
+  @protected
+  @mustCallSuper
+  void handleAppLifecycleStateChanged(ui.AppLifecycleState state) {
+    _lifecycleState = state;
+  }
+
+  /// Initializes the [lifecycleState] with the [initialLifecycleState] from the
+  /// window.
+  ///
+  /// Once the [lifecycleState] is populated through any means (including this
+  /// method), this method will do nothing. This is because the
+  /// [initialLifecycleState] may already be stale and it no longer makes sense
+  /// to use the initial state at dart vm startup as the current state anymore.
+  ///
+  /// The latest state should be obtained by subscribing to
+  /// [WidgetsBindingObserver.didChangeAppLifecycleState].
+  @protected
+  void readInitialLifecycleStateFromNativeWindow() {
+    if (_lifecycleState == null && _parseAppLifecycleMessage(window.initialLifecycleState) != null) {
+      _handleLifecycleMessage(window.initialLifecycleState);
+    }
+  }
+
+  Future<String> _handleLifecycleMessage(String message) async {
+    handleAppLifecycleStateChanged(_parseAppLifecycleMessage(message));
+    return null;
+  }
+
+  static ui.AppLifecycleState _parseAppLifecycleMessage(String message) {
+    switch (message) {
+      case 'AppLifecycleState.paused':
+        return ui.AppLifecycleState.paused;
+      case 'AppLifecycleState.resumed':
+        return ui.AppLifecycleState.resumed;
+      case 'AppLifecycleState.inactive':
+        return ui.AppLifecycleState.inactive;
+      case 'AppLifecycleState.detached':
+        return ui.AppLifecycleState.detached;
+    }
+    return null;
   }
 }
